+++ src/components/User/UserAbout/style.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const UserAboutNotFound = styled.div`
+  text-align: center;
+  color: #868e96;
+  margin-top: 10rem;
+  font-size: 2.5rem;
+`;
